--- src/api/graphql/index.ts
@@ -1,5 +1,0 @@
-import { userResolver, albumResolver } from "./resolvers/index.js";
-import { userGQLSchema, albumGQLSchema } from "./schema/index.js";
-
-export const typeDefs = [userGQLSchema, albumGQLSchema];
-export const resolvers = [userResolver, albumResolver];
