--- src/api/graphql/resolvers/index.ts
@@ -1,2 +1,0 @@
-export * from "./userResolver.js";
-export * from "./albumResolver.js";
