--- src/api/graphql/schema/index.ts
@@ -1,2 +1,0 @@
-export * from "./userSchema.js";
-export * from "./albumSchema.js";
