+++ src/graphql/resolvers/index.ts
@@ -1,0 +1,2 @@
+export * from "./userResolver.js";
+export * from "./albumsResolver.js";
