+++ src/graphql/resolvers/userResolver.ts
@@ -1,5 +1,6 @@
 import axios from "axios";
-import { IUser } from "../../../features/user/models/User.js";
+import { IUser } from "../../models/User.js";
+import { SPOTIFY_BASE } from "../../lib/constants/spotify.js";
 
 export const userResolver = {
   Query: {
@@ -7,7 +8,7 @@ export const userResolver = {
       if (!user) {
         throw new Error("User not found");
       }
-      const result = await axios.get("https://api.spotify.com/v1/me", {
+      const result = await axios.get(`${SPOTIFY_BASE}/v1/me`, {
         headers: { Authorization: `Bearer ${user.spotifyToken.accessToken}` },
       });
