+++ src/graphql/schema/index.ts
@@ -1,0 +1,2 @@
+export * from "./userSchema.js";
+export * from "./albumsSchema.js";
